fixing incorrect handling of operands
When fetching an instruction, we need to read the
second and third words (if present) *before* executing
the instruction. GetOperand and SetOperand() coupled the
reading and decoding of operands together with
setting and getting of operands, which is wrong.

Instead, we introduce an Operand class, with several
subclasses. Each instruction is fully decoded before
execution.