refactor DCPU insn decode to be stateless
Instead of modifying the CPU state during an
instruction fetch and decode, we update variables representing
PC and SP. This allows us to decode an arbitrary number of
instructions ahead of time (e.g. for debug output) without
affecting program execution. It also fixes a bug with IF*
instructions where the SP might be incorrectly modified when
skipping the next instruction.

Also deleted most of the operand-related tests, as they're
becoming tedious to maintain and their purpose has largely been
served.